rename of cache service from initialize to factory
